refactor(entities-api): standardize api module conventions
