+++ apps/web/src/entities/auth/api/auth.queries.ts
@@ -1,6 +1,6 @@
 import { createQueryKeys } from '@lukemorales/query-key-factory'
 
-import { authGet } from './auth-get'
+import { authGet } from './auth.api'
 
 export const authQueries = createQueryKeys('auth', {
   getUser: () => ({
