+++ apps/web/src/entities/fcm/api/fcm.api.ts
@@ -1,9 +1,10 @@
 import { api } from '@/shared/lib/ky'
+
 import type { FcmPostResponseDTO } from '@/entities/fcm/model/fcm.type'
 
 export const fcmPost = {
   postFcmToken: async (token: string): Promise<FcmPostResponseDTO> => {
-    const response = await api.post(`fcm/register-token`, { json: { token } })
+    const response = await api.post('fcm/register-token', { json: { token } })
 
     return {
       ok: response.ok,
