+++ apps/web/src/entities/fcm/api/index.ts
@@ -1,0 +1,1 @@
+export * from './fcm.api'
