+++ apps/web/src/entities/image/api/index.ts
@@ -1,1 +1,1 @@
-export * from './image-api'
+export * from './image.api'
