+++ apps/web/src/entities/mypage/api/alert.api.ts
@@ -1,0 +1,22 @@
+import { api } from '@/shared/lib/ky'
+
+import type { Alert, AlertResponseDTO } from '@/entities/mypage/model/alert.type'
+
+export const alertGet = {
+  getAlert: async (): Promise<AlertResponseDTO> => {
+    const response = await api.get('mypage/alert').json<AlertResponseDTO>()
+    return response
+  },
+}
+
+export const alertPatch = {
+  patchAlert: async ({
+    startAlert,
+    inGameAlert,
+  }: Alert): Promise<AlertResponseDTO> => {
+    const response = await api
+      .patch('mypage/alert', { json: { startAlert, inGameAlert } })
+      .json<AlertResponseDTO>()
+    return response
+  },
+}
